LayoutTests/imported/w3c/web-platform-tests/offscreen-canvas/drawing-rectangles-to-the-canvas/2d.clearRect.nonfinite-expected.txt

2d.clearRect.nonfinite

clearRect() with Infinity/NaN is ignored


PASS clearRect() with Infinity/NaN is ignored 
